[#173]refactor:  회원가입 폴더 구조 변경
