+++ src/pages/Authentication/Join/Join1/Styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import * as I from '../../../assets/Icons';
+import * as I from '../../../../assets/Icons';
 
 export const JoinLayout = styled.div`
   width: 100%;
