+++ src/pages/Authentication/Join/Join3/Styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { JoinLayout,LogoContainer,ZimeetLogo, LogoTitle, LogoText, JoinText} from '../Join/Styles';
+import { JoinLayout,LogoContainer,ZimeetLogo, LogoTitle, LogoText, JoinText} from '../Join1/Styles';
 
 export { JoinLayout, LogoContainer, ZimeetLogo, LogoTitle, LogoText, JoinText};
